Revert "Revert "Revert "Merge pull request #5 from primitivefinance/release/hotfix-missing-interfaces"""
This reverts commit 997120588e4ea568559576d94d7b12d75dd174e5.

+++ packages/primitive-contracts/contracts/Primitives.sol
@@ -1,12 +1,13 @@
 // SPDX-License-Identifier: MIT
+
 pragma solidity ^0.6.2;
 
 library Primitives {
     struct Option {
         address underlyingToken;
         address strikeToken;
-        uint256 base;
-        uint256 quote;
-        uint256 expiry;
+        uint base;
+        uint quote;
+        uint expiry;
     }
 }
